+++ src/main/java/org/kumoricon/site/attendee/PrintBadgeHandler.java
@@ -7,7 +7,7 @@ import org.kumoricon.model.attendee.Attendee;
 import java.util.List;
 
 public interface PrintBadgeHandler {
-    void showAttendeeBadgeWindow(AttendeePrintView view, List<Attendee> attendeeList);
+    void showAttendeeBadgeWindow(AttendeePrintView view, List<Attendee> attendeeList, boolean forcePrintAll);
     void badgePrintSuccess(PrintBadgeWindow printBadgeWindow, List<Attendee> attendees);
     void reprintBadges(PrintBadgeWindow printBadgeWindow, List<Attendee> attendeeList);
     BadgePrintFormatter getBadgeFormatter(PrintBadgeWindow printBadgeWindow, List<Attendee> attendees);